Updated DB schema, to better accomodate uint64
This is to work around a weird problem where pgx would get mixed up
with uint64 numbers near their upper range, and bogusly not be able
to insert them *sometimes*.

+++ database/schema.sql
@@ -2,8 +2,8 @@
 -- PostgreSQL database dump
 --
 
--- Dumped from database version 10.8
--- Dumped by pg_dump version 10.8
+-- Dumped from database version 10.9
+-- Dumped by pg_dump version 10.9
 
 SET statement_timeout = 0;
 SET lock_timeout = 0;
@@ -81,7 +81,7 @@ CREATE TABLE public.execution_run (
     module_name text,
     program_counter bigint,
     stack_top bigint,
-    value bigint,
+    value numeric,
     preserve_top boolean,
     discard integer,
     condition_met boolean,
